Fix crash caused by UsbTransfer finalization logic error
One of the cardinal rules for finalizers is this: Don't try to access any
objects (structs are fine) via the 'this' pointer, because they may have
already been finalized!  This is the reason for the "disposing" parameter
to the Dispose(disposing) pattern -- to inform code whether it's OK to try
accessing other objects to clean them up.

UsbTransfer's Dispose(disposing) implementation disregards the 'disposing'
parameter and unconditionally tries to do a full cleanup of the transfer.
However, if the dependent objects are finalized first, they will throw an
exception on the finalizer thread when the Dispose method tries to use them.

This patch fixes this issue.