Allow DeviceNotifyEventArgs to be subclassed in client code
Addresses #18.

These changes should allow clients to write their own subclasses of
DeviceNotifyEventArgs, which should be possible as this class is
referenced within the interface IDeviceNotifier.

The fact that IDeviceNotifier is publicly exposed suggests that
third-party users should be able to write their own implementations of
this interface, which they cannot easily do if they cannot properly
subclass DeviceNotifyEventArgs.