Use NativeLibrary.SetDllImportResolver to resolve native libraries at runtime
The name of the native library (libusb-1.0) can vary from platform to platform.

This PR updates the project to use NativeLibrary.SetDllImportResolver to resolve native libraries at runtime. This method is available in .NET Core 3.0 and above.

This avoids having to build multiple versions of the .NET library for multiple platforms, which was done using the third-party MSBuild.Sdk.Extras project.

There's no need to be explicity about the calling convention; .NET Core will default to StdCall on Windows and Cdecl on all other platforms.