Fix access violation issue (#5)
This is a weird bug. The mMonoTransferContext member of
MonoUsbEndPointReader and MonoUsbEndPointWriter is never used and
instead the transfer context is always stored in the mTransferContext
member of UsbEndpointBase. This mTransferContext is never disposed of properly
and instead GC will dispose of it when it destroys the transfer context
objects. Destroying the transfer context objects involves a
dereferencing opearation to libusb, and this can happen after libusb_exit()
has been called. At this point the internal structures of libusb are
all destroyed and a access violation exception is inevitable.

(The Ignore SYSLOB0011 commit is needed to make the project compile for dotnet versions > 6.0)